Use correct graphql type for secondsToWait (#20655)
Summary:
This can be a float, so reflect that in the schema (or it will raise a
type error when a non-int is used here). Unused field in frontend
currently.

Test Plan: BK

## Test Plan

## Changelog

> If a changelog entry is required, replace this section with the
desired entry.
>
> By default, no changelog entry will be produced for internal PRs.

Synced-From-Internal

GitOrigin-RevId: 4b14b720a8fc24290be75f65543417fec97b2db8

+++ python_modules/dagster-graphql/dagster_graphql/schema/logs/events.py
@@ -171,7 +171,7 @@ class GrapheneExecutionStepRestartEvent(graphene.ObjectType):
 
 
 class GrapheneExecutionStepUpForRetryEvent(graphene.ObjectType):
-    secondsToWait = graphene.Field(graphene.Int)
+    secondsToWait = graphene.Field(graphene.Float)
 
     class Meta:
         interfaces = (GrapheneMessageEvent, GrapheneStepEvent, GrapheneErrorEvent)
